Removed rake 0.8.7 dependency specific to ruby 1.9.2, as it didn't seem to break anything without it
Made border definitions access instead of delete hash params, as calling
it from within a loop without cloning the style hash was causing it to
be empty on subsequent loops.  Didn't see any reason to be deleting
them.

Added :edges param to :border hash to specify [:top,:bottom], etc.  If
it's absent, defaults to all edges.