Fix worksheet name escaping and cell sorting in Axlsx.cell_range
We've discovered that Excel do not allow &apos; in defined_names to
reference a worksheet name. All it does allow is to escape the simple
quote by doubling it. Given that Worksheet#name is escaped using
HTMLEntities, a simple quote appears as &apos; we're simply replacing it by
the correct '' when using absolute reference to worksheet range.