Update bg_color in conditional formatting
Conditional formatting for background color fills is different for DXF.  This must be undocumented and is quite annoying.  I verified the behavior in Excel 2010 Mac and Windows.  I didn't understand why background colors were not being applied even though the style was being applied for conditional formatting.  Looking at styles.xml I saw that it is different in a file created by Excel.  I have updated the code to reflect it and tested in Mac and Windows.