Integrate workbook views and alter serialization
This integrates workbook views and sheet state into serialization.

I also noticed that we were populating defined names during
serialization. While it is good to delay this as late as possible
as there is always the chance that some conditional programming
by the consumer adds, and then removes an autofilter, I am
choosing to risk it at this point for cleaner code.