Do not put Axlsx::DefinedName#name in serialized_attributes (fixes #285)
When using serialized_attributes both attribute's name and value are camelcased, in
`Axlsx::DefinedName` the values starts with `_xmln` which is then
transformed in `Xmln`.

The fix proposed on #285 does not work because then other `Axlsx::DefinedName`
attributes name are no more camelcased (`localSheetId` become
`local_sheet_id` for instance).

As proposed by @randym in #285, this commit only make a special case of
the name and let the other attributes go through serialized_attributes.

Adding a test on it revealed a wrong test which would have catched
similar error but was not because of value returned by `doc.xpath`.