Add auto_filter defined name in `Axlsx::Worksheet#to_xml_string` (refs #285)
The way ot was done previously (adding the defined name on
`Axlsx::Worksheet#auto_filter=`) meant that it was added only when
using this method. But it's possible to assign range on the auto-created
`AutoFilter` with:

```ruby
workheet.auto_filter.range = "A1:B2"
```

In this case the defined name was never added to the workbook.