Axlsx.sanitize uses delete() vs. delete!() for frozen strings
Whenever a frozen string is passed as an input to any sanitized value, we are modifying it in place which raised a RuntimeError if that string is frozen (as you might expect constants like header or workbook names to be).  Use the safer delete() method which creates a new, modified copy of the string.